goit-de-hw-06/requirements.txt
pyspark>=3.5.0
kafka-python
